// frontend/src/api/generated/model/notificationResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { NotificationResponseType } from './notificationResponseType';

export interface NotificationResponse {
  id: number;
  type: NotificationResponseType;
  content: string;
  isRead: boolean;
  relatedPostId?: number;
  relatedCommentId?: number;
  createdAt: string;
}